add support for recent glslify
fixes #6

The most recent version of glslify (6.1.0) requires that transforms
export a sync interface. I have left the async interface as is, for
compatiblity with older versions.

I have upgrade the tests to run with the newest version of glslify, but
before doing that, I asserted that the old tests still pass with an old
glslify version.

I have also bumped the version in package.json to 3.1.0, increasing the
minor version because this is not a breaking change.